+++ CHANGELOG.md
@@ -4,9 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+
 ## [Unreleased]
 
 ### Added
+
+### Fixed
+
+
+## [0.1.0] - 2019-04-11
+
+First beta version.
 
 ### Fixed